Emit metrics
Use the decorated Producer and Consumer interfaces in http-gateway and function-sidecar, respectively.

Note: gather metrics on the function input topic only and not the function output or reply topics.

Producers are identified in producer metrics by a string. Currently a UUID is used.

Pods are identified in consumer metrics by a string. Currently a UUID is used.
This can be changed later to some more compact identifier such as pod name
(and namespace) or pod IP address.