Support the emission of transport metrics
Decorate the Producer and Consumer interfaces with metrics-emitting
equivalents. Provide a MetricsReceiver interface and implementation.

Apply the decorations to the Kafka producer and consumer and emit metrics to
Kafka (hence the "kafka_over_kafka" package name). Provide a metrics receiver
constructor function which uses a Kafka consumer to receive metrics.

Metrics aggregation is not yet implemented. Aggregate metrics with a count of
1 and a duration of 0 are emitted.