Consume metrics and use them for auto-scaling
Avoid scaling down to zero when the input topic is non-empty. This is
necessary because production and consumption rates could be zero for a while
and yet the input topic could contain arbitrarily many unprocessed messages.

The autoscaler unit tests resort to white box testing to avoid races. Rather
than running receive and sampling loops, the autoscaler is told when to:
  * receive specific metrics, and
  * take a sample.

The autoscaler is a collection of decorated metrics scalers indexed by
function id. Each metrics scaler proposes a number of replicas based on
metrics. The decorations:
  * limit scaling up according to configuration,
  * limit scaling down according to input topic queue length, and
  * delay scaling to zero for a configurable time period.